Updated to generic views and django forms
Updated form.py to include a constructor that returns the fields needed in the form.

Removed the html and added django form in questionnaire template

Still to be done:
Updated views.py to only have generic classes, the QuizView uses 'FormView 'and 'DetailView' as it needs the get_object function to pass to the template. This may need to be changed.

Not all of the urls in views.py are currently named urls.

AddView currently does not take the user to the newly created page, it just takes them back.