Updated FormView to include all questions from the quiz
Updated forms.py to include a for loop iterating through each quiz question displaying the question and the four options as radio buttons.

Updated view.py to include a list that holds the choices of the user from the form, passing this information into the result.html

Updated results.html to be a for loop that compares each questions actual answer to the answer the user put.

Added crispy forms to the templates and removed boostrapting, currently only has basic crisp as form helpers gets repeated for each for loop.

Fixed the buttons not working for create/edit templates.

Added some list filters to the django admin.

Removed two templates as they are not needed now with the for loops.