Added jquery to display one question at a time
Added animation for the questions to display one by one.
Shows the correct answer if the user got it wrong.
Now show the score/amount of correct answers and allows the user to retry the quiz with the same button.

Results view and template can be removed as is not longer used.

The html to show each question is currently badly implemented with repeat code  and only working for quizzies with less than 5 questions.